Move `mise` initialization after other path mods.
This is a bit of a chicken and egg thing, we need `~/.local/bin` to be
earlier in `$PATH` for the default installation location of `mise`, but
it also means that anything `mise` installs will override the "local"
bin folders that normally are last in line.

There's probably a better solution to this but I'm content to see if
having `mise` up front on `$PATH` actually causes problems before doing
anything about it.

Incidentally, this didn't cause problems previously because I had an
error in something in `/etc/profile.d/` that caused this file to be
double-sourced, slowing down all my logins on my home server for years.
Wheee...`